add html and obj to viewObj and don't add html to view element if obj is valid
add the response to the view object in html, and if the response is JSON, parse it and set it to the view object obj property. If the response is valid json then the user will not want to display it, so don't set it to the innerHTML. Providing the html in the object also simplifies things for the using programmer.